+++ priv/components/badge.exs
@@ -21,7 +21,7 @@
       size: ["extra_small", "small", "medium", "large", "extra_large"],
       rounded: ["extra_small", "small", "medium", "large", "extra_large", "full", "none"],
       only: ["badge"],
-      helpers: [hide_badge: 2, show_badge: 2],
+      helpers: [hide_badge: 1, hide_badge: 2, show_badge: 1, show_badge: 2],
       module: ""
     ],
     optional: [],