+++ priv/components/banner.exs
@@ -23,7 +23,7 @@
       rounded: ["extra_small", "small", "medium", "large", "extra_large", "none"],
       padding: ["extra_small", "small", "medium", "large", "extra_large", "none"],
       only: ["banner"],
-      helpers: [show_banner: 2, hide_banner: 2],
+      helpers: [show_banner: 1, show_banner: 2, hide_banner: 1, hide_banner: 2],
       module: ""
     ],
     optional: [],