+++ priv/components/drawer.exs
@@ -20,7 +20,7 @@
       ],
       size: ["extra_small", "small", "medium", "large", "extra_large"],
       only: ["drawer"],
-      helpers: [hide_drawer: 3, show_drawer: 3],
+      helpers: [hide_drawer: 2, hide_drawer: 3, show_drawer: 2, show_drawer: 3],
       module: ""
     ],
     optional: [],