+++ priv/components/tabs.exs
@@ -19,7 +19,7 @@
       size: ["extra_small", "small", "medium", "large", "extra_large"],
       padding: ["extra_small", "small", "medium", "large", "none"],
       only: ["tabs"],
-      helpers: [show_tab: 3, hide_tab: 3],
+      helpers: [show_tab: 2, show_tab: 3, hide_tab: 2, hide_tab: 3],
       module: ""
     ],
     optional: [],