+++ priv/components/toast.exs
@@ -23,7 +23,7 @@
       rounded: ["extra_small", "small", "medium", "large", "extra_large", "none"],
       type: ["toast", "toast_group"],
       only: ["toast", "toast_group"],
-      helpers: [show_toast: 2, hide_toast: 2],
+      helpers: [show_toast: 1, show_toast: 2, hide_toast: 1, hide_toast: 2],
       module: ""
     ],
     optional: [],